requirements.txt
docutils==0.12
lockfile==0.12.2
python-daemon==2.1.1
wheel==0.24.0
